Refactor payroll system: update employee role labels, streamline payroll rejection process, and enhance payroll report functionality
- Changed role labels from "Staff" to "Pegawai" across various components.
- Introduced RejectPayrollModal for handling payroll rejection with reason input.
- Updated payroll detail and list pages to accommodate rejection status and actions.
- Removed unnecessary conditions and fields related to employee attributes.
- Adjusted payroll report to display only paid statuses and improved UI elements.
- Cleaned up unused functions and variables in the payroll condition preview logic.

+++ backend/database/migrations/2026_06_06_113558_create_phase3_tables.php
@@ -16,13 +16,13 @@ return new class extends Migration
             $table->foreignId('employee_id')->constrained()->cascadeOnDelete();
             $table->char('period_month', 7); // e.g., 2026-06
             
-            $table->decimal('wfo_days', 8, 2)->default(0);
-            $table->decimal('wfh_days', 8, 2)->default(0);
-            $table->decimal('out_of_town_days', 8, 2)->default(0);
+            $table->unsignedInteger('wfo_days')->default(0);
+            $table->unsignedInteger('wfh_days')->default(0);
+            $table->unsignedInteger('out_of_town_days')->default(0);
             $table->integer('business_trips')->default(0);
-            $table->decimal('training_days', 8, 2)->default(0);
-            $table->decimal('overtime_hours', 8, 2)->default(0);
-            $table->decimal('total_mandays', 8, 2)->default(0);
+            $table->unsignedInteger('training_days')->default(0);
+            $table->unsignedInteger('overtime_hours')->default(0);
+            $table->unsignedInteger('total_mandays')->default(0);
             $table->unsignedInteger('late_count')->default(0);
             
             $table->boolean('is_finalized')->default(false);
